requirements.txt
google-generativeai
langchain-google-genai
langgraph
python-dotenv
fpdf
langchain
ffmpeg
jsonschema